+++ src/main/java/org/SurvivalOutbreak/PlayerLocation.java
@@ -1,0 +1,7 @@
+package org.SurvivalOutbreak;
+
+import com.github.hanyaeger.api.Coordinate2D;
+
+public interface PlayerLocation {
+    Coordinate2D getLocation();
+}
