+++ src/main/java/org/SurvivalOutbreak/ui/AmmoText.java
@@ -1,0 +1,28 @@
+package org.SurvivalOutbreak.ui;
+
+import com.github.hanyaeger.api.Coordinate2D;
+import com.github.hanyaeger.api.entities.impl.TextEntity;
+import javafx.scene.paint.Color;
+import javafx.scene.text.Font;
+import javafx.scene.text.FontWeight;
+
+public class AmmoText extends TextEntity {
+
+    public AmmoText(Coordinate2D initialLocation) {
+        super(initialLocation);
+        setFont(Font.font("Roboto", FontWeight.BOLD, 20));
+        setFill(Color.WHITE);
+    }
+
+    public void update(int currentAmmo, int maxAmmo) {
+        setText("Ammo: " + currentAmmo + " / " + maxAmmo);
+
+        if (currentAmmo == 0) {
+            setFill(Color.RED);
+        } else if (currentAmmo <= maxAmmo * 0.25) {
+            setFill(Color.ORANGE);
+        } else {
+            setFill(Color.WHITE);
+        }
+    }
+}
