+++ src/main/java/org/SurvivalOutbreak/ui/ScoreText.java
@@ -1,0 +1,4 @@
+package org.SurvivalOutbreak.ui;
+
+public class ScoreText {
+}
