--- src/main/java/org/example/PlayerLocation.java
@@ -1,7 +1,0 @@
-package org.example;
-
-import com.github.hanyaeger.api.Coordinate2D;
-
-public interface PlayerLocation {
-    Coordinate2D getLocation();
-}
