--- src/main/java/org/example/ui/ScoreText.java
@@ -1,4 +1,0 @@
-package org.example.ui;
-
-public class ScoreText {
-}
